backend/requirements.txt
# Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
ultralytics==8.0.200
pygame==2.5.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
